third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-transition-events-expected.txt

Verifies that pointerenter/leave/over/out are fired correctly for correponding mouse events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


green received mouseover, at phase=target
grey received mouseover, at phase=bubble
grey received mouseenter, at phase=target
green received mouseenter, at phase=target
green received mouseout, at phase=target
grey received mouseout, at phase=bubble
green received mouseleave, at phase=target
grey received mouseleave, at phase=target
PASS successfullyParsed is true

TEST COMPLETE
